# project/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Custom Bikes",
        "contact": {
            "name": "Developer Contact",
            "url": "https://www.example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/marchandises": {
            "get": {
                "tags": [
                    "marchandise"
                ],
                "summary": "List marchandise",
                "responses": {
                    "200": {
                        "description": "Success: List all marchandise",
                        "schema": {
                            "$ref": "#/definitions/Marchandise"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "marchandise"
                ],
                "summary": "Create Marchandise",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Marchandise"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success: A Newly Created Marchandise",
                        "schema": {
                            "$ref": "#/definitions/Marchandise"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            }
        },
        "/api/marchandises/{id}": {
            "get": {
                "tags": [
                    "marchandise"
                ],
                "summary": "Get Marchandise by Id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Display the specified Marchandise by id.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Marchandise",
                        "schema": {
                            "$ref": "#/definitions/Marchandise"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            },
            "put": {
                "tags": [
                    "marchandise"
                ],
                "summary": "Update Marchandise",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Update the specified Marchandise by id.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Marchandise"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Marchandise updated",
                        "schema": {
                            "$ref": "#/definitions/Marchandise"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            },
            "delete": {
                "tags": [
                    "marchandise"
                ],
                "summary": "Delete Marchandise",
                "description": "Delete the specified Marchandise by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Marchandise id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    },
                    "204": {
                        "description": "Success: successful deleted"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Marchandise": {
            "required": [
                "libelle"
            ],
            "properties": {
                "libelle": {
                    "description": "libelle d\u00e9crivant la marchandise",
                    "type": "string",
                    "example": "Malta"
                },
                "type_marchandise_id": {
                    "description": "type d\u00e9crivant la marchandise",
                    "type": "string",
                    "example": "1"
                }
            }
        },
...
    }
}